requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
uuid==1.30
Werkzeug==3.0.3